backend/requirements.txt
fastapi
uvicorn
httpx==0.27.0
python-dotenv==1.0.1
itsdangerous==2.2.0
sqlite-utils==3.36
pydantic
google-api-python-client
reportlab
APScheduler
yagmail